# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: FlowState API
  description: A minimal, standalone RESTful API scaffold for backend development.
  version: 0.1.0
paths:
...
  /api/checkpoints:
    get:
      tags:
      - checkpoints
      summary: List Checkpoints
      operationId: list_checkpoints_api_checkpoints_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: all
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FrontendCheckpoint'
                title: Response List Checkpoints Api Checkpoints Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/checkpoints/{checkpoint_id}/approve:
    post:
      tags:
      - checkpoints
      summary: Approve Checkpoint
      operationId: approve_checkpoint_api_checkpoints__checkpoint_id__approve_post
      parameters:
      - name: checkpoint_id
        in: path
        required: true
        schema:
          type: string
          title: Checkpoint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendCheckpoint'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/checkpoints/{checkpoint_id}/confirm-deps:
    post:
      tags:
      - checkpoints
      summary: Confirm Deps Checkpoint
      description: 'Testing stage Phase 1: confirm dep installation and trigger Phase
        2 (install + run tests).'
      operationId: confirm_deps_checkpoint_api_checkpoints__checkpoint_id__confirm_deps_post
      parameters:
      - name: checkpoint_id
        in: path
        required: true
        schema:
          type: string
          title: Checkpoint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendCheckpoint'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/checkpoints/{checkpoint_id}/reject:
    post:
      tags:
      - checkpoints
      summary: Reject Checkpoint
      operationId: reject_checkpoint_api_checkpoints__checkpoint_id__reject_post
      parameters:
      - name: checkpoint_id
        in: path
        required: true
        schema:
          type: string
          title: Checkpoint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectCheckpointRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontendCheckpoint'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FrontendCheckpoint:
      properties:
        id:
          type: string
          title: Id
        pipelineId:
          type: string
          title: Pipelineid
        pipelineName:
          type: string
          title: Pipelinename
        stage:
          type: string
          title: Stage
        stageIndex:
          type: integer
          title: Stageindex
        status:
          type: string
          title: Status
        createdAt:
          type: string
          title: Createdat
        output:
          type: string
          title: Output
        rejectReason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejectreason
        subPhase:
          anyOf:
          - type: string
          - type: 'null'
          title: Subphase
        depsManifest:
          anyOf:
          - type: object
          - type: 'null'
          title: Depsmanifest
        reviewScore:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reviewscore
        reviewIssues:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Reviewissues
        passRate:
          anyOf:
          - type: string
          - type: 'null'
          title: Passrate
      type: object
      required:
      - id
      - pipelineId
      - pipelineName
      - stage
      - stageIndex
      - status
      - createdAt
      - output
      title: FrontendCheckpoint
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RejectCheckpointRequest:
      properties:
        reason:
          type: string
          title: Reason
      type: object
      required:
      - reason
      title: RejectCheckpointRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
servers:
- url: http://127.0.0.1:8000
  description: Local development server
- url: http://localhost:8000
  description: Alternate local development server
...